neuro_brush/requirements.txt
django
djangorestframework
torch
torchvision
pillow
matplotlib
ipython
